Symlinks now also works on embedded root filesystems relative to the system rootfs.
Now it's possible to fix and change symlinks in embedded root filesystems (e.g. /gpfs/sys/ubuntu11.04) without the need of chroot or fakechroot.
Therefore option -e has been added, to tell symlinks that it operates in a relative rootfs.
The current working directory is taken as the relative path to the system root, so symlinks has to be called within the rootfs directory.